+++ src/sisl/io/siesta/_help.py
@@ -98,129 +98,7 @@ def _csr_from(col_from, csr):
     csr.translate_columns(col_from, col_to)
 
 
-def _mat2dtype(M, dtype: np.dtype) -> None:
-    """Change the internal CSR matrix in `M` to a follow `dtype`"""
-
-    if M.dtype == dtype:
-        return M
-
-    spin = M.spin
-    csr = M._csr
-    shape = csr._D.shape
-
-    # Change details
-    old_dtype = np.dtype(M.dtype)
-    new_dtype = np.dtype(dtype)
-
-    def toc(D, re, im):
-        return (D[..., re] + 1j * D[..., im]).astype(dtype, copy=False)
-
-    if old_dtype.kind in ("f", "i"):
-        if new_dtype.kind in ("f", "i"):
-            # this is just simple casting
-            csr._D = csr._D.astype(dtype)
-        elif new_dtype.kind == "c":
-            # we need to *collect it
-            if spin.is_diagonal:
-                # this is just simple casting,
-                # each diagonal component has its own index
-                csr._D = csr._D.astype(dtype)
-            elif spin.is_noncolinear:
-                D = np.empty(shape[:-1] + (shape[-1] - 1,), dtype=dtype)
-                # These should be real only anyways!
-                D[..., [0, 1]] = csr._D[..., [0, 1]].real.astype(dtype)
-                D[..., 2] = toc(csr._D, 2, 3)
-                if D.shape[-1] > 4:
-                    D[..., 3:] = csr._D[..., 4:].astype(dtype)
-                csr._D = D
-            elif spin.is_spinorbit:
-                D = np.empty(shape[:-1] + (shape[-1] - 4,), dtype=dtype)
-                D[..., 0] = toc(csr._D, 0, 4)
-                D[..., 1] = toc(csr._D, 1, 5)
-                D[..., 2] = toc(csr._D, 2, 3)
-                D[..., 3] = toc(csr._D, 6, 7)
-                if D.shape[-1] > 4:
-                    D[..., 4:] = csr._D[..., 8:].astype(dtype)
-                csr._D = D
-            elif spin.is_nambu:
-                D = np.empty(shape[:-1] + (shape[-1] - 8,), dtype=dtype)
-                D[..., 0] = toc(csr._D, 0, 4)
-                D[..., 1] = toc(csr._D, 1, 5)
-                D[..., 2] = toc(csr._D, 2, 3)
-                D[..., 3] = toc(csr._D, 6, 7)
-                D[..., 4] = toc(csr._D, 8, 9)  # S
-                D[..., 5] = toc(csr._D, 10, 11)  # Tuu
-                D[..., 6] = toc(csr._D, 12, 13)  # Tdd
-                D[..., 7] = toc(csr._D, 14, 15)  # T0
-                if D.shape[-1] > 8:
-                    D[..., 8:] = csr._D[..., 16:].astype(dtype)
-                csr._D = D
-            else:
-                raise NotImplementedError
-        else:
-            raise NotImplementedError
-
-    elif old_dtype.kind == "c":
-        if new_dtype.kind == "c":
-            # this is just simple casting
-            csr._D = csr._D.astype(dtype)
-        elif new_dtype.kind in ("f", "i"):
-            # we need to *collect it
-            if spin.is_diagonal:
-                # this is just simple casting,
-                # each diagonal component has its own index
-                csr._D = csr._D.astype(dtype)
-            elif spin.is_noncolinear:
-                D = np.empty(shape[:-1] + (shape[-1] + 1,), dtype=dtype)
-                # These should be real only anyways!
-                D[..., [0, 1]] = csr._D[..., [0, 1]].real.astype(dtype)
-                D[..., 2] = csr._D[..., 2].real.astype(dtype)
-                D[..., 3] = csr._D[..., 2].imag.astype(dtype)
-                if D.shape[-1] > 4:
-                    D[..., 4:] = csr._D[..., 3:].real.astype(dtype)
-                csr._D = D
-            elif spin.is_spinorbit:
-                D = np.empty(shape[:-1] + (shape[-1] + 4,), dtype=dtype)
-                D[..., 0] = csr._D[..., 0].real.astype(dtype)
-                D[..., 1] = csr._D[..., 1].real.astype(dtype)
-                D[..., 2] = csr._D[..., 2].real.astype(dtype)
-                D[..., 3] = csr._D[..., 2].imag.astype(dtype)
-                D[..., 4] = csr._D[..., 0].imag.astype(dtype)
-                D[..., 5] = csr._D[..., 1].imag.astype(dtype)
-                D[..., 6] = csr._D[..., 3].real.astype(dtype)
-                D[..., 7] = csr._D[..., 3].imag.astype(dtype)
-                if D.shape[-1] > 8:
-                    D[..., 8:] = csr._D[..., 4:].real.astype(dtype)
-                csr._D = D
-            elif spin.is_nambu:
-                D = np.empty(shape[:-1] + (shape[-1] + 8,), dtype=dtype)
-                D[..., 0] = csr._D[..., 0].real.astype(dtype)
-                D[..., 1] = csr._D[..., 1].real.astype(dtype)
-                D[..., 2] = csr._D[..., 2].real.astype(dtype)
-                D[..., 3] = csr._D[..., 2].imag.astype(dtype)
-                D[..., 4] = csr._D[..., 0].imag.astype(dtype)
-                D[..., 5] = csr._D[..., 1].imag.astype(dtype)
-                D[..., 6] = csr._D[..., 3].real.astype(dtype)
-                D[..., 7] = csr._D[..., 3].imag.astype(dtype)
-                D[..., 8] = csr._D[..., 4].real.astype(dtype)  # S
-                D[..., 9] = csr._D[..., 4].imag.astype(dtype)
-                D[..., 10] = csr._D[..., 5].real.astype(dtype)  # Tuu
-                D[..., 11] = csr._D[..., 5].imag.astype(dtype)
-                D[..., 12] = csr._D[..., 6].real.astype(dtype)  # Tdd
-                D[..., 13] = csr._D[..., 6].imag.astype(dtype)
-                D[..., 14] = csr._D[..., 7].real.astype(dtype)  # T0
-                D[..., 15] = csr._D[..., 7].imag.astype(dtype)
-                if D.shape[-1] > 16:
-                    D[..., 16:] = csr._D[..., 8:].real.astype(dtype)
-                csr._D = D
-            else:
-                raise NotImplementedError
-        else:
-            raise NotImplementedError
-    M._reset()
-
-
-def _mat_siesta2sisl(M, dtype: Optional[np.dtype] = None) -> None:
+def _mat_siesta2sisl(M) -> None:
     """Conversion of Siesta spin matrices to sisl spin matrices
 
     The matrices from Siesta are given in a format adheering to the following
@@ -266,9 +144,6 @@ def _mat_siesta2sisl(M, dtype: Optional[np.dtype] = None) -> None:
     On top of this it depends on whether the data-type is complex
     or not.
     """
-    if dtype is None:
-        dtype = M.dtype
-
     spin = M.spin
 
     if spin.kind in (spin.NONCOLINEAR, spin.SPINORBIT, spin.NAMBU):
@@ -277,17 +152,9 @@ def _mat_siesta2sisl(M, dtype: Optional[np.dtype] = None) -> None:
         else:
             M._csr._D[:, 2] = M._csr._D[:, 2].conj()
 
-    _mat2dtype(M, dtype)
 
-
-def _mat_sisl2siesta(M, dtype: Optional[np.dtype] = None) -> None:
+def _mat_sisl2siesta(M) -> None:
     """Conversion of sisl to Siesta spin matrices"""
-    if dtype is None:
-        dtype = M.dtype
-
-    # convert to float
-    _mat2dtype(M, dtype)
-
     spin = M.spin
 
     if spin.kind in (spin.NONCOLINEAR, spin.SPINORBIT, spin.NAMBU):