+++ src/sisl/physics/tests/test_physics_sparse.py
@@ -234,6 +234,34 @@ def test_sparse_orbital_bz_spin_orbit(dtype):
     assert np.abs((MT - MH)._csr._D).sum() != 0
 
 
+def test_sparse_orbital_bz_spin_orbit_astype():
+    Mr = SparseOrbitalBZSpin(geom.graphene(), spin=Spin("SO"), dtype=np.float64)
+    Mc = SparseOrbitalBZSpin(geom.graphene(), spin=Spin("SO"), dtype=np.complex128)
+
+    p0 = [0.1, 0.2, 0.3, 0.4, 0.0, 0.0, 0.3, -0.4]
+    p1 = [0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9]
+
+    Mr.construct(
+        (
+            [0.1, 1.44],
+            [p0, p1],
+        )
+    )
+    p0 = _so_real2cmplx(p0)
+    p1 = _so_real2cmplx(p1)
+    Mc.construct(
+        (
+            [0.1, 1.44],
+            [p0, p1],
+        )
+    )
+
+    assert np.allclose(Mr.astype(np.complex128)._csr._D, Mc._csr._D)
+    assert np.allclose(Mc.astype(np.float64)._csr._D, Mr._csr._D)
+    assert np.allclose(Mr.astype(np.complex128).astype(np.float64)._csr._D, Mr._csr._D)
+    assert np.allclose(Mc.astype(np.float64).astype(np.complex128)._csr._D, Mc._csr._D)
+
+
 def _nambu_cmplx2real(p):
     return [
         p[0].real,
@@ -301,6 +329,55 @@ def test_sparse_orbital_bz_nambu(dtype):
     assert np.abs((M - MT)._csr._D).sum() != 0
     assert np.abs((M - MH)._csr._D).sum() == 0
     assert np.abs((MT - MH)._csr._D).sum() != 0
+
+
+def test_sparse_orbital_bz_nambu_astype():
+    Mr = SparseOrbitalBZSpin(geom.graphene(), spin=Spin("nambu"), dtype=np.float64)
+    Mc = SparseOrbitalBZSpin(geom.graphene(), spin=Spin("nambu"), dtype=np.complex128)
+
+    p0 = [
+        0.1 + 1j * 0.0,
+        0.2 + 1j * 0.0,
+        0.3 + 1j * 0.4,
+        0.3 - 1j * 0.4,
+        # onsite S must have zero real
+        # onsite triplet states must have 0 imaginary
+        1j * 0.6,
+        0.3,
+        0.4,
+        0.3,
+    ]
+
+    p1 = [
+        0.2 + 1j * 0.6,
+        0.3 + 1j * 0.7,
+        0.4 + 1j * 0.5,
+        0.3 + 1j * 0.9,
+        0.3 + 1j * 0.7,
+        0.4 + 1j * 0.8,
+        0.5 + 1j * 0.6,
+        0.4 + 1j * 1.0,
+    ]
+
+    Mc.construct(
+        (
+            [0.1, 1.44],
+            [p0, p1],
+        )
+    )
+    p0 = _nambu_cmplx2real(p0)
+    p1 = _nambu_cmplx2real(p1)
+    Mr.construct(
+        (
+            [0.1, 1.44],
+            [p0, p1],
+        )
+    )
+
+    assert np.allclose(Mr.astype(np.complex128)._csr._D, Mc._csr._D)
+    assert np.allclose(Mc.astype(np.float64)._csr._D, Mr._csr._D)
+    assert np.allclose(Mr.astype(np.complex128).astype(np.float64)._csr._D, Mr._csr._D)
+    assert np.allclose(Mc.astype(np.float64).astype(np.complex128)._csr._D, Mc._csr._D)
 
 
 @pytest.mark.filterwarnings("ignore", message="*is NOT Hermitian for on-site")
@@ -376,11 +453,12 @@ def test_sparse_orbital_transform_ortho_unpolarized():
     assert np.abs(Mt.tocsr(2)).sum() == 0
     assert np.abs(Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[0] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mt.tocsr(-1)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[0] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mt.tocsr(-1)).sum() == 0
 
 
 def test_sparse_orbital_transform_nonortho_unpolarized():
@@ -405,11 +483,12 @@ def test_sparse_orbital_transform_nonortho_unpolarized():
     assert np.abs(Mt.tocsr(2)).sum() == 0
     assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[0] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[0] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
 
 def test_sparse_orbital_transform_ortho_polarized():
@@ -432,11 +511,12 @@ def test_sparse_orbital_transform_ortho_polarized():
     assert np.abs(Mt.tocsr(2)).sum() == 0
     assert np.abs(Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mt.tocsr(-1)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mt.tocsr(-1)).sum() == 0
 
 
 def test_sparse_orbital_transform_ortho_nc():
@@ -459,11 +539,12 @@ def test_sparse_orbital_transform_ortho_nc():
     assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
     assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
 
 
 @pytest.mark.filterwarnings("ignore", message="*is NOT Hermitian for on-site")
@@ -487,11 +568,12 @@ def test_sparse_orbital_transform_ortho_so():
     assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
     assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
 
 
 @pytest.mark.filterwarnings("ignore", message="*is NOT Hermitian for on-site")
@@ -518,12 +600,13 @@ def test_sparse_orbital_transform_nonortho_so():
     assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
     assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="so")
-    assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
-    assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
-    assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
-    assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
-    assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
+    for s in ("so", "nambu"):
+        Mt = M.transform(spin=s)
+        assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
+        assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
+        assert np.abs(Mcsr[2] - Mt.tocsr(2)).sum() == 0
+        assert np.abs(Mcsr[3] - Mt.tocsr(3)).sum() == 0
+        assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
 
 def test_sparse_orbital_transform_basis():
@@ -551,32 +634,42 @@ def test_sparse_orbital_transform_combinations():
     M.finalize()
     Mcsr = [M.tocsr(i) for i in range(M.shape[2])]
 
-    Mt = M.transform(spin="non-colinear", dtype=np.float64, orthogonal=True).transform(
-        spin="polarized", orthogonal=False
+    Mt = (
+        M.transform(spin="non-colinear", orthogonal=True)
+        .astype(np.float64)
+        .transform(spin="polarized", orthogonal=False)
     )
     assert M.dim == Mt.dim
     assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
     assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
     assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(dtype=np.float128, orthogonal=True).transform(
-        spin="so", dtype=np.float64, orthogonal=False
+    Mt = (
+        M.transform(orthogonal=True)
+        .astype(np.float128)
+        .transform(spin="so", orthogonal=False)
+        .astype(np.float64)
     )
     assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
     assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
     assert np.abs(Mt.tocsr(2)).sum() == 0
     assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="polarized", orthogonal=True).transform(
-        spin="so", dtype=np.float64, orthogonal=False
+    Mt = (
+        M.transform(spin="polarized", orthogonal=True)
+        .transform(spin="so", orthogonal=False)
+        .astype(np.float64)
     )
     assert np.abs(Mcsr[0] - Mt.tocsr(0)).sum() == 0
     assert np.abs(Mcsr[1] - Mt.tocsr(1)).sum() == 0
     assert np.abs(Mt.tocsr(2)).sum() == 0
     assert np.abs(Mcsr[-1] - Mt.tocsr(-1)).sum() == 0
 
-    Mt = M.transform(spin="unpolarized", dtype=np.float32, orthogonal=True).transform(
-        dtype=np.complex128, orthogonal=False
+    Mt = (
+        M.transform(spin="unpolarized", orthogonal=True)
+        .astype(np.float32)
+        .transform(orthogonal=False)
+        .astype(np.complex128)
     )
     assert np.abs(0.5 * Mcsr[0] + 0.5 * Mcsr[1] - Mt.tocsr(0)).sum() == 0
 
@@ -597,27 +690,23 @@ def test_sparse_orbital_transform_matrix():
     assert Mt.dim == 2
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(1)).sum() == 0
 
-    Mt = M.transform(matrix=np.ones((3, 3)), dtype=np.float64)
+    Mt = M.transform(matrix=np.ones((3, 3)))
     assert Mt.dim == 3
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(2)).sum() == 0
 
-    Mt = M.transform(
-        spin="non-colinear", matrix=np.ones((4, 3)), orthogonal=True, dtype=np.float64
-    )
+    Mt = M.transform(spin="non-colinear", matrix=np.ones((4, 3)), orthogonal=True)
     assert Mt.dim == 4
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(3)).sum() == 0
 
-    Mt = M.transform(spin="non-colinear", matrix=np.ones((5, 3)), dtype=np.float64)
+    Mt = M.transform(spin="non-colinear", matrix=np.ones((5, 3)))
     assert Mt.dim == 5
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(4)).sum() == 0
 
-    Mt = M.transform(
-        spin="so", matrix=np.ones((8, 3)), orthogonal=True, dtype=np.float64
-    )
+    Mt = M.transform(spin="so", matrix=np.ones((8, 3)), orthogonal=True)
     assert Mt.dim == 8
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(7)).sum() == 0
 
-    Mt = M.transform(spin="so", matrix=np.ones((9, 3)), dtype=np.float64)
+    Mt = M.transform(spin="so", matrix=np.ones((9, 3)))
     assert Mt.dim == 9
     assert np.abs(Mcsr[0] + Mcsr[1] + Mcsr[2] - Mt.tocsr(8)).sum() == 0